e2e enable bzip2 and add test for bzip2 decompression.
Note that we can only decompress bzip2, not compress to bzip2. I didn't
see any places in the current code where this would cause an issue, but
it's a bit awkward. Can you think of any issues or how we should deal
with this differently?

Although our bzip2 implementation does not have CRC checking, that's
probably fine as openpgp has integrity protection in other places.