Fix overflow e2e.incrementalByteArray which affects OTR CTR mode.
When pass Uint8Array to incrementalByteArray, the "carry" does not work,  because when n[i] = 255, we have n[i] += 1 (= 0), so carry = (n[i] & 0x100) >>> 8 ( = 0). As a consequence, when we use it in OTR CTR mode [1], the counter never propagates to the next byte and we use repeated counter over and over again.

[1] Fortunately, our main CTR mode use ByteArray which is not overflowed.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=120456836