Bring external codebase in sync with google3 version.
Unfortunately, both versions have diverged significantly and are essentially two independent forks at this point, so it's not going to be possible to manually tag every change individually.

Going forward google3 will be the only source of truth, and any pull requests will be resolved internally on the google3 version, as that one is significantly more maintained and used.

Summary:
 - Remove goog.isDef/Null.
 - Remove extraProvide supressions.
 - Use ES6 classes rather than Closure classes.
 - Add missing require/provides.
 - Blind exponents on ElGamal and RSA.
 - Formatting and readability updates.
 - File permissions changes.
 - 25519 curve maintenance changes.